pyproject.toml
[project]
name = "addressprinter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-wtf>=1.2.2",
    "pillow>=12.0.0",
    "pyap>=0.3.1",
    "pylabels>=1.2.1",
    "reportlab>=4.4.6",
]

[dependency-groups]
prod = [
    "gunicorn>=23.0.0",
]
